Update TRAM packaging.
It seems that an update to setuptools has broken our ability to package
TRAM, e.g. running the docker build results in "“No module named 'tram'”
errors. Upon investigation, setuptools was creating all of the install
metadata for the package (including the `tram` executable) but not
actually installing the tram package. My fix is to require
setuptools>=61 and move the setup.cfg contents to pyproject.toml. This
is fixing the docker build in my testing.